Add feedback toasts
Added toast messages to provide feedback on operations like create and
delete. The toasts are created using a ngx-toast library that is a
wrapper for well established 'toastr' library.

Other changes:

Provided utility callback to modal service that is only called, when the
user's response is 'yes' (to remove need to always verify the result in
function code).