Add session restoration
Added logic that resumes session based on auth token stored into the
local storage when the page is relaoded.

Note:

This sample project uses basic auth (and possibly over unsecure http
connection), which is not secure and can leak the authentication token
(which can be then decoded to username + password). Therefore, it should
not be used in any actual production environment.

It would be better to provide e.g. some opaque authorization token from
the server (or use Oauth2) that has expiration time, but this project is
for understanding Angular concepts and not concentrating too much on
proper security/authentication aspects.