Add lazy-loading of modules
Added lazy loading mechanism to load the feature modules (auth, recipes,
shopping lists) lazily when the module related view is accessed for the
first time.

This was done for a practise reason as the modules are not yet that
large that it would make much sense to optimize the loading mechanism
here.