Use $ suffix in observable names
Modified all observable class property and input parameter names to
include '$' character at the end of the name when the value type is
'Observable'. This convention is suggested in e.g. some of the
Angular documentation's guides and it is especially useful when the
ngrx based selectors (that returns observables) are used in fields that
are referenced in the template (i.e. it's easier to know when the
'async' pipe method is used).

It also provides nice distinction to the observables compared to more
tranditional value types (primitives, etc).