Fix issues in authentication
Fixed two issues related to authentication.

First, the authentication related event had API error catching on outer
piped scope that lead into "killing" the action observable when API
returned an error (e.g. user mistyped password on the first attempt).
This lead the app staying in loading state when re-trying login.

Second, the form data was reset when the login/signup form was reset
(without waiting the result). As there is a loading indicator that
blocks any user input for the form while the authentication operation is
on going, there is no need to clear the form. But when the
authentication fails on any reason, it is more user-friendly if the form
content remains the same so that the user can e.g. re-apply the password
without needing to type the username again, etc.

Now the form "state" remains until user is redirected after successful
login/signup.