Refactor: improve API response error structuring
Refactored some API services to use wrapping mechanism to return api
specific errors in a wrapped interface that is easy to work with in
TypeScript. In addition, it works well along with the ngrx effects
compared to the model where the errors are catched and re-thrown on api
service level. In that model it was sometimes hard to get the
observables work as expected.

Other changes:

Modified header component to utilize a bit more generalized utility
function to provide feedback on recipes data loading/storing operation.
It should be now more easy to use the same pattern on different
components.