Include Angular PWA / service worker
Included @angular/pwa module with service worker and configured it to
provided caching strategy for recipes and their ingredients. This way
when the server is offline, the app can still work at some degraded
level and the recipes are vievable.

This was done just to experience how well the Angular PWA works and what
can be achieved easily with it.