Fix hover issue on recipe list item
Fixed a small visual issue on recipe list item related to the case when
an item was both selected (clicked) and the mouse is hovering over it.
It looked a bit silly that the hover styling was applied to the selected
item.

Now the hover effect does not affect selected item (other than having a
pointer cursor).