Fix issue in new recipe button press
Fixed a small issue related to clicking 'new recipe' button on recipe
list when the current route is in "edit recipe". Previously the included
"deselect recipe" action caused the navigation to just navigate up one
level (to recipe details) and it required a second press to the button
to get "edit recipe" view for a new recipe.

Now one button press is enough to allow navigating into new recipe
route.